[Fix] 🔧 DeliveryAddressDTO 자료형 수정

+++ GoGuMa/src/main/java/com/ggm/goguma/dto/DeliveryAddressDTO.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class DeliveryAddressDTO {
-	private int addressId;
+	private long addressId;
 	private String nickName;
 	private String recipient;
 	private String address;
 	private String contact;
 	private int isDefault;
-	private int memberId;
+	private long memberId;
 }
